% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_duration_curves-GEN.R
\name{fdc}
\alias{fdc}
\alias{fdc_compute}
\alias{fdc_plot}
\title{Flow duration curves}
\usage{
fdc_compute(cosdata)

fdc_plot(cos_data, log_y = TRUE, log_x = FALSE, ...)
}
\arguments{
\item{cosdata}{A data.frame with columns as used throughout visCOS}
}
\description{
Flow duration curves

Computes the flow duration curves (fdc) for the \emph{cosdata} \code{tibble}.
The calculations are adapted from the method used within the hydroTSM package.

Plots the flow duration curves (fdc) for the \emph{cosdata} \code{tibble}.
The function uses \code{ggplot} to so and facets the different basins into
separate subplots. Each subplot shows the fdc of the \eqn{o}-data and
the \eqn{s}-data.
}
