% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge-GEN.R, R/judge_explore-GEN.R
\name{judge}
\alias{judge}
\alias{judge_explore}
\title{Wrapper for the different of functions}
\usage{
judge(cosdata, type = "compute", of_metrics = list(nse = coscos::of_nse, 
    kge = coscos::of_kge, pbias = coscos::of_pbias, corr = coscos::of_cor), 
    opts = coscos::viscos_options(), ...)

judge_explore(cos_data, of_metrics = list(nse = coscos::of_nse, kge = coscos::of_kge, 
    pbias = coscos::of_pbias, corr = coscos::of_cor), opts = coscos::viscos_options())
}
\description{
This function provides a wrapper for all `judge_` functions. The
input arguments `type` is lazyily-evaluated (which basically implies that
it does not have to be provided as a character argument).
which function is chosen (the "pre fix" part of
the selected functions, i.e. `judge_` and `of_`, do not need to be provided).
Th

Runs a Shiny Gadget which can be used to get an overview of a cos_data time
series object.
}
\examples{
# get example data, explore the model performance
cos_data <- get_viscos_example()
judge_explore(cos_data)
}
