% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_compare-GEN.R, R/judge_explore-GEN.R
\name{judge_explore2}
\alias{judge_explore2}
\alias{judge_explore}
\title{Explore with Objective Functions}
\usage{
judge_explore2(d1, d2 = NULL, of_metrics = list(nse = coscos::of_nse, 
    kge = coscos::of_kge, pbias = coscos::of_pbias, corr = coscos::of_cor), 
    opts = coscos::viscos_options())

judge_explore(cosdata, .ofuns = list(nse = coscos::of_nse, kge = coscos::of_kge, 
    pbias = coscos::of_pbias, corr = coscos::of_cor), opts = coscos::viscos_options())
}
\description{
Runs a Shiny Gadget which can be used to interactively explore two 
`cos_data` objects in terms of their respective objective functions.

Runs a Shiny Gadget which can be used to get an overview of a cosdata time 
series object.
}
\seealso{
Other judge functions: \code{\link{judge}}
}
