% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge-GEN.R
\name{judge_barplot}
\alias{judge_barplot}
\alias{judge_rasterplot}
\title{Bar plot for the Main Objective Function Values}
\usage{
judge_barplot(cosdata, .ofuns = list(nse = coscos::of_nse, kge = coscos::of_kge, 
    pbias = coscos::of_pbias, corr = coscos::of_cor))

judge_rasterplot(cosdata, .ofuns = list(nse = coscos::of_nse, kge = coscos::of_kge, 
    pbias = coscos::of_pbias, corr = coscos::of_cor))
}
\description{
Bar plot for the Main Objective Function Values

Bar plot for the Main Objective Function Values
}
