% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_window-GEN.R
\name{judge_window}
\alias{judge_window}
\title{Objectie Function Windowing}
\usage{
judge_window(cosdata, of_metrics = list(nse = coscos::of_nse, kge = coscos::of_kge, 
    pbias = coscos::of_pbias, corr = coscos::of_cor), window_sizes = c(20L, 500L), 
    na_filling = FALSE, opts = coscos::viscos_options())
}
\description{
Compute the objective function for a set of windows.
}
