% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/of-GEN.R
\name{of}
\alias{of}
\title{Objective Functions Wrap}
\usage{
of(cos_data, d_metrics = list(nse = d_nse, kge = d_kge, pbias = d_pbias, corr = d_cor), 
    type = "compute", ...)
}
\description{
Objective Functions Wrap
}
