% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peack-GEN.R
\name{peack}
\alias{peack}
\title{Event-wise Exploration of the high-runoffs}
\usage{
peack(cosdata, action = "compute", n_events = 10L, window_size = 24L, 
    opts = coscos::viscos_options())
}
\description{
Event-wise Exploration of the high-runoffs
}
