% Please edit documentation in R/of-GEN.R
\name{plot_main_of}
\alias{plot_main_of}
\title{Plot main objective function values}
\description{
Currently two options for plotting the main objectives are provided by
visCOS: Plotting the different objective functions values as a set of
bar plots \code{barplot_of} and plotting a summary table in form of
a large raster of all the objective function values \code{rasterplot_of}.
}
