% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_window-GEN.R, R/windowing-GEN.R
\name{roll_roll}
\alias{roll_roll}
\alias{roll_roll}
\title{Objectie Function Rolling Functionallity}
\usage{
roll_roll(d1, d2 = NULL, fun, width, ...)

roll_roll(d1, d2 = NULL, fun, width, ...)
}
\description{
Roll an objective fucntion over two sets of data (d1 and d2)

Roll an objective fucntion over two sets of data d1 and 2
}
\keyword{internal}
