% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_window-GEN.R
\name{window_fun}
\alias{window_fun}
\title{Objectie Function Windowing}
\usage{
window_fun(cos_data, of_metric, window_sizes = c(20L, 500L), na_filling = FALSE, 
    opts = coscos::viscos_options())
}
\description{
Compute the objective function for a set of windows.
}
