% Please edit documentation in R/windowing-GEN.R
\name{window_of}
\alias{window_of}
\title{Objectie Function Windowing}
\usage{
window_of(cos_data, ofun, lb = 0, window_sizes = c(20L, 500L), na_filling = FALSE, 
    opts = coscos::viscos_options())
}
\description{
Compute the objective function for a set of windows.
}
