% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_aggregation-GEN.R
\name{aggregate_time}
\alias{aggregate_time}
\title{Time Aggregation}
\usage{
aggregate_time(cosdata, key = "mm", .funs = base::mean, opts = coscos::viscos_options())
}
\arguments{
\item{cosdata}{Data-Format used within visCOS. A longer definition and explanation is provided in the introduction vignette; and the \code{coscos} function \pkg{cook_cosdata} provides a pre-set function to transform and check if your data is in the correct format).}
}
\description{
Aggregates \code{cosdata}) according to the
timely resolution defined via \code{aggregation}. Possible
resolution-choices are \code{'yyyy'} - year, \code{'mm'} - month and
\code{'dd'} - day and combinations thereof.
}
