% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge-GEN.R, R/judge_explore-GEN.R
\name{judge}
\alias{judge}
\alias{judge_explore}
\title{Wrapper for the different judge functions}
\usage{
judge(cosdata, type = "compute", .ofuns = list(nse = coscos::of_nse, 
    kge = coscos::of_kge, pbias = coscos::of_pbias, corr = coscos::of_cor), 
    opts = coscos::viscos_options(), ...)

judge_explore(cosdata, .ofuns = list(nse = coscos::of_nse, kge = coscos::of_kge, 
    pbias = coscos::of_pbias, corr = coscos::of_cor), opts = coscos::viscos_options())
}
\description{
A wrapper for all `judge_` functions. The
input arguments `type` is lazily-evaluated (which basically implies that
it does not have to be provided as a character argument).

Runs a Shiny Gadget which can be used to get an overview of a cosdata time
series object.
}
\examples{
# get example data, explore the model performance
cosdata <- get_viscos_example()
judge_explore(cosdata)
}
