% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge-GEN.R
\name{judge}
\alias{judge}
\alias{judge_compute}
\alias{judge_barplot}
\alias{judge_rasterplot}
\title{Wrapper for several judge-functions}
\usage{
judge(cosdata, type = "compute", .ofuns = list(nse = coscos::of_nse, 
    kge = coscos::of_kge, pbias = coscos::of_pbias, corr = coscos::of_cor), 
    opts = coscos::viscos_options(), ...)

judge_compute(cosdata, .ofuns = list(nse = coscos::of_nse, kge = coscos::of_kge, 
    pbias = coscos::of_pbias, corr = coscos::of_cor), opts = coscos::viscos_options())

judge_barplot(cosdata, .ofuns = list(nse = coscos::of_nse, kge = coscos::of_kge, 
    pbias = coscos::of_pbias, corr = coscos::of_cor))

judge_rasterplot(cosdata, .ofuns = list(nse = coscos::of_nse, kge = coscos::of_kge, 
    pbias = coscos::of_pbias, corr = coscos::of_cor))
}
\arguments{
\item{cosdata}{cosdata data.frame.}
}
\value{
list of basic objective function evaluated for the different
hydrological years and over the whole timespan.
}
\description{
A wrapper for several \code{judge} functions. The input arguments \code{type} is lazily-evaluated (that basically means that it does not have to be provided as a character argument).

Computes one or several objective functions for each basin in the \emph{cosdata} \code{tibble}.

Bar plot for the Main Objective Function Values

Bar plot for the Main Objective Function Values
}
\seealso{
Other judge functions: \code{\link{judge_explore2}}
}
