% Please edit documentation in R/of-GEN.R
\name{judge}
\alias{judge}
\title{Wrapper for the different of functions}
\usage{
judge(cos_data, type = "compute", d_metrics = list(nse = d_nse, kge = d_kge, 
    pbias = d_pbias, corr = d_cor), ...)
}
\description{
This function provides a wrapper for all `judge_` functions.
}
