% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge-GEN.R
\name{judge_compute}
\alias{judge_compute}
\title{Get basic objective function for}
\usage{
judge_compute(cosdata, .ofuns = list(nse = coscos::of_nse, kge = coscos::of_kge, 
    pbias = coscos::of_pbias, corr = coscos::of_cor), opts = coscos::viscos_options())
}
\arguments{
\item{cosdata}{cosdata data.frame.}
}
\value{
list of basic objective function evaluated for the different
hydrological years and over the whole timespan.
}
\description{
Calculate basic objective functions(NSE, KGE, percentage BIAS, Correlation)
for every basin and the chosen periods.
}
