% Please edit documentation in R/of-GEN.R
\name{judge_barplot}
\alias{judge_barplot}
\alias{judge_rasterplot}
\title{Bar plot for the Main Objective Function Values}
\usage{
judge_barplot(cosdata, d_metrics = list(nse = d_nse, kge = d_kge, pbias = d_pbias, 
    corr = d_cor))

judge_rasterplot(cos_data, d_metrics = list(nse = d_nse, kge = d_kge, pbias = d_pbias, 
    corr = d_cor))
}
\description{
Bar plot for the Main Objective Function Values

Bar plot for the Main Objective Function Values
}
